refactor: mv durable agent with agent
