+++ dapr_agents/agents/__init__.py
@@ -1,3 +1,3 @@
 from .base import AgentBase
 from .agent.agent import Agent
-from dapr_agents.workflow.agents.durable import DurableAgent
+from .durableagent.agent import DurableAgent
