+++ dapr_agents/workflow/__init__.py
@@ -2,5 +2,4 @@ from .base import WorkflowApp
 from .task import WorkflowTask
 from .agentic import AgenticWorkflow
 from .orchestrators import LLMOrchestrator, RandomOrchestrator, RoundRobinOrchestrator
-from .agents import DurableAgent
 from .decorators import workflow, task
